made latency and packet size static arguments
Before latency was expressed as send duration and recv duration but that
is kinda complicated and confusing and no the focus of the project, so
all send latencies are the same and messages will not get dropped  if
they arrive too late. packet size is the largest message you're able to
send, but the buffer size is different because it can contain multiple
packets. If packets were larger than buffers than no messages would be
received